Refactor code actions to use generators, and respect .only
From the [docs](https://microsoft.github.io/language-server-protocol/specifications/specification-current/#textDocument_codeAction):

> Requested kind of actions to return.
>
> Actions not of this kind are filtered out by the client before
> being shown, so servers can omit computing them.

This means we can skip the tsserver protocol request when only is
present but doesn't contain the type of action we're trying to find.

This is a little complicated by the fact that code action kinds aren't
necessarily simple equality checks - they're dot-chained strings. So,
if the client is only processing `refactor` we still need `refactor.move`
and vice versa. I'm comparing on the most basic prefixes, so this should
work.