Upgrade vscode-languageserver (#231)
This upgrades the core vscode-langaugeserver package and the LSP in use to v3.16.0. Relates to #227 (comment).

Also pulled in vscode-languageserver-textdocument. The language server's TextDocument has been deprecated and replaced
with this package.

Spent a lot of time getting the example dir working. Because the monaco language server package is not maintained much
any more, it's incompatible with the newer version of the language server protocol (even when upgraded) and that leads
to needing to pin different versions. I've attempted a few different upgrades and hit various typescript and webpack
errors along the way. I ended up deciding to just get the example working (and added some better instructions about how
to verify this) and skip upgrading anything inside it.