Lookup tsserver using direct path rather than through .bin alias (#234)
Lookup the tsserver script through a more explicit typescript/lib/tsserver.js
path rather than through the alias created in the ".bin" directory. This
in theory is less future proof as the path to tsserver could change but
in practice it likely never will. And we also had other code paths
(yarn sdk lookup or lookup within this server's package) that used that
approach already so this unifies things more.

The side-effect of doing that is that we'll use "child_process.fork"
rather than "child_process.spawn" for starting the tsserver but we've
done that in other cases already and it seems like a better approach
anyway as then ENV will be inherited from the parent process (haven't
really verified that).

Also it fixes a corner case like #217 where "tsc" package is used and
makes the server fail.

Relates to #217