Implement handling for the APPLY_COMPLETION_CODE_ACTION command
Until now, the APPLY_COMPLETION_CODE_ACTION had to be implemented by
each client separately.

There are two parts to the APPLY_COMPLETION_CODE_ACTION command. Its
arguments include both the "command" and "changes" properties.

The "changes" property contains workspace edits and this can be provided
by the server itself and this is what this PR does.

The "command" property is an arbitrary typescript command that this PR
doesn't handle but neither client can really handle it since it needs
to be resolved against tsserver. Since I haven't seen and don't know of
any use case for handling "command", this is ignored in this PR. If
clients know how to handle those arbitrary commands (which is doubtful),
then they can still handle the APPLY_COMPLETION_CODE_ACTION command
instead of relying on this implementation.