fix: apply path normalized both to paths both from client and server
This fixes issues with mixed paths on Windows that would result
in server file paths not matching the client file paths and some
functionality to misfunction.

The client can pass URIs in various formats:
 - file://c:/folder/file.ts
 - file://C:/folder/file.ts
 - file://c%3A/folder/file.ts

So both the case of the drive letter can differ and the colon might or
might not be encoded.

As for the file paths returned from the server there can be:
 - C:/folder/file.ts
 - c:/folder/file.ts
 - C:\folder\file.ts

So both the drive letter case can and the path separator can differ.

When we are getting URI from the client, we convert it to file path
and store in map. When the server returns a file path from some API
result then that path might not match what we got from the client. Note
that the server can auto-discover files itself and that's where the
different path formats are coming from.

Fix by normalizing both the client paths before storing them in the map
and the server paths before returning to the client or comparing with
client paths.

 The colon part is not really relevant for this
 issue but as for the