fix: normalize client and tsserver paths (#275)
This fixes issues with mixed paths on Windows that would result
in server file paths not matching the client file paths and some
functionality to misfunction.

The client can pass URIs in various formats:

    file://c:/folder/file.ts
    file://C:/folder/file.ts
    file://c%3A/folder/file.ts

So both the case of the drive letter can differ and the colon might or
might not be encoded.

As for the file paths returned from the tsserver there can be:

    C:/folder/file.ts
    c:/folder/file.ts
    C:\folder\file.ts

So both the drive letter case and the path separator can differ.

When we are getting URI from the client, we are converting it to a file
path and storing in a map. When the server returns a file path in
some API result then that path might not match what we've gotten from
the client. (Note that the tsserver can auto-discover files itself and
that's where the different path formats are coming from.)

Fix by normalizing both the client paths before storing them in the map
and the server paths before returning to the client or comparing with
client paths.

BTW. The difference in colon encoding should not really be an issue
since this is already normalized when converting from an URI.