feat: Include import specifier when showing completions
For completions that import from another package, the completions
will include a "detail" field with the name of the module.

This works only with the "allowIncompleteCompletions" typescript option
enabled so did that by default.

Also aligned some other options and logic with the typescript language
services used in VSCode:
 - enabled various other typescript options that are enabled there
 - annotating completions with the local name of the import when
   completing the path in `import foo from '...'`
 - update completion `sortText` regardless if the completion "isRecommended"

Fixes #280