fix: add semver as a dependency
The `semver` package is used in the source code, but is not declared as
a dependency in `package.json`. This causes errors when trying to run
`typescript-language-server`, because the `semver` package is missing.

In development, it is installed as a transitive dependency of
`@typescript-eslint` tooling. Those dev dependencies are not installed
in production, hence the problem.