fix: remove support for spawning the tsserver through shell script
The support for passing path to a `tsserver` shell script or just
a `tsserver` string to `--tsserver-path`, and relying it to be found
on the PATH, is removed.

The provided path must point to the `typescript/lib` directory (or
`typescript/lib/tsserver.js` script for backward compatibility) and the
server will run it through the `tsserver.js` path that it will look up.

Starting `tsserver` through a shell script is wrong and in some cases
even causes server to not work at all. That's because spawning
a node process for `tsserver` instead of forking the current one can
result in running a completely different node version or even a node
version that is sandboxed and can't communicate with this LSP server.

Fixes #191